feat: split Governance screen in two.
- separate governance/plasma and governance/token screens
- navigation in submenu
