+++ src/routes/governance/plasma.tsx
@@ -7,17 +7,16 @@
 
 import * as React from 'react';
 import { observer } from 'mobx-react';
-import { List, Collapse, Icon, Divider, Row } from 'antd';
+import { List, Collapse, Icon } from 'antd';
 import TimeAgo from 'react-timeago';
 
-import Web3SubmitWarning from '../components/web3SubmitWarning';
-import TokenGovernance from './tokenGovernance';
+import Web3SubmitWarning from '../../components/web3SubmitWarning';
 
-import AppLayout from '../components/appLayout';
-import EtherscanLink from '../components/etherscanLink';
-import { shortenHex } from '../utils';
+import AppLayout from '../../components/appLayout';
+import EtherscanLink from '../../components/etherscanLink';
+import { shortenHex } from '../../utils';
 import { Button } from 'antd';
-import { governanceContractStore } from '../stores/governanceContract';
+import { governanceContractStore } from '../../stores/governanceContract';
 
 const { Fragment } = React;
 
@@ -130,55 +129,47 @@ export default class Governance extends React.Component<GovernanceProps, any> {
       finalize,
     } = governanceContract;
     return (
-      <AppLayout section="governance">
-        <Row gutter={16}>
-          <h1>Minimum Viable Governance</h1>
-          <Web3SubmitWarning />
+      <AppLayout section="governance/plasma">
+        <h1>Leap Network Governance</h1>
+        <Web3SubmitWarning />
 
-          {noGovernance && <span>No governance set up</span>}
+        {noGovernance && <span>No governance set up</span>}
 
-          {proposals && proposals.length === 0 && <span>No proposals yet</span>}
-          {proposals && proposals.length > 0 && (
-            <Fragment>
-              <h1>Proposals</h1>
-              <List
-                itemLayout="vertical"
-                size="small"
-                dataSource={proposals}
-                renderItem={item => (
-                  <List.Item
-                    className="governanceProposal"
-                    extra={
-                      <div
-                        style={{
-                          minWidth: '190px',
-                          textAlign: 'center',
-                          padding: '16px',
-                        }}
-                      >
-                        {renderDate(item)}
-                      </div>
-                    }
-                  >
-                    <List.Item.Meta title={item.summaryStr} />
-                    {this.renderContent(item)}
-                  </List.Item>
-                )}
-              />
-            </Fragment>
-          )}
-          {canFinalize && (
-            <div style={{ paddingTop: '1rem' }}>
-              <Button onClick={finalize}>Finalize proposals</Button>
-            </div>
-          )}
-        </Row>
-        <Row>
-          <Divider />
-        </Row>
-        <Row gutter={16}>
-          <TokenGovernance />
-        </Row>
+        {proposals && proposals.length === 0 && <span>No proposals yet</span>}
+        {proposals && proposals.length > 0 && (
+          <Fragment>
+            <h1>Proposals</h1>
+            <List
+              itemLayout="vertical"
+              size="small"
+              dataSource={proposals}
+              renderItem={item => (
+                <List.Item
+                  className="governanceProposal"
+                  extra={
+                    <div
+                      style={{
+                        minWidth: '190px',
+                        textAlign: 'center',
+                        padding: '16px',
+                      }}
+                    >
+                      {renderDate(item)}
+                    </div>
+                  }
+                >
+                  <List.Item.Meta title={item.summaryStr} />
+                  {this.renderContent(item)}
+                </List.Item>
+              )}
+            />
+          </Fragment>
+        )}
+        {canFinalize && (
+          <div style={{ paddingTop: '1rem' }}>
+            <Button onClick={finalize}>Finalize proposals</Button>
+          </div>
+        )}
       </AppLayout>
     );
   }
